fixes for some warnings with gcc 4.4
git-svn-id: https://svn.berlios.de/svnroot/repos/vermont/trunk/vermont@2169 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ src/modules/idmef/IDMEFMessage.cpp
@@ -181,7 +181,7 @@ void IDMEFMessage::applyVariables()
 		string key = string("%") + iter->first + "%";
 		string val = iter->second;
 
-		uint32_t pos;
+		size_t pos;
 		while ((pos = message.find(key)) != string::npos) {
 			message.replace(pos, key.size(), val);
 		}
